Able to update the status of each task
The user can type done along with the task number separated by a space to update the status of the task as "done".

Minor changes in the skeleton to suit the changes in this new commit.

Messages will be printed if the task number is invalid or if the input is "list", if there are no tasks in the list