Code complies with the given coding standard
From this commit onwards, every commit will have a body.

So far, from Level-1 to Level-3, this chatbot creates a list where user can save his tasks and mark "done" if they have completed it. They can also see the list of tasks after increments. Messages will be printed when the user commands the bot to print out a list if there are no tasks or if the task number given by the user to mark as done is out of range.