Fix bug where the done command is executed erroneously
If user's input contained the word "done", even in the middle, the
done command would initiate.

This would prevent the user from typing any task that had the word
"done" in it.

Now, user input is being read in and parsed for the command, which
is stored in userCommand.

This gives the user assurance that the "done" command will only be
initiated if the user types it at the front of his input.