Add semi-automated testing.
Added automated testing for command line UI.
Moved creation of scanner object out of the loop in main() so that it
only gets created once. Originally was inside the getUserInput() method,
causing it to be created every loop. This caused the automated testing
to only read one line of the input.txt file.

Slight modifications to List, Deadline and Event classes to account for
the extra " " between the task description and "/by" or "/at".