Categorising the tasks into three types
The most important change in this commit is that the name of the bot has been changed to Austin for uniqueness.

Now, there are three different types of tasks: todo, event and deadline. Todo is a type of task with no dates. Event is a type of task which contains the date of the event. Deadline is a type of task containing the date and time of the task deadline.

There is a new command called "help" which prints out all the commands along with its description and format. This is used whenever the user is unsure of any command.

Another change is that leading and trailing spaces of the input command is removed, so that the command is executed though there are additional spaces between the words and at the start of the command.