A-TextUiTesting
Reformatted some of the output text.
Made tiny changes to variable names.
Made changes such that the chat bot can handle description with a whitespace " ".
Removed any non-ASCII characters in the text art to prevent the runtest.bat from throwing errors.
Added test cases and expected outcome to test all the commands.
Added a copy of the text-art folder into the text-ui-test as the working directory of that is different.
All test case passed...