Refactor command handling and output methods to reduce complexity
The current class containing the main method contains many other
methods. It is also deeply nested due to commands being handled in the
main method.

Moving the command handling and the methods into separate classes would
help to reduce the complexity of the code, which makes maintaining the
code and fixing bugs easier.

Let's
* move command handling from main to a new class
* move message output methods from main to another new class