Increment Extension A-Exceptions
Added the DukeException Class and its constructor. Implemented the unusual cases to trigger DukeException error messages. Also, made some changes to main function to be more abstract and cleaner, and some changes to the passing of parameters. Lastly, a small change to the naming convention to error message when todo name is empty.