Merge branch-Level-5
The error handlings are changed to try-catch statements to make the error handling easier. This is done by creating new exception class called AustinException, which is thrown when the command is invalid or when the format of the command is incorrect.

The format for event and deadline commands are changed, so that it will be easier to type in. Instead of "/by" and "/at", "|" symbol is used to differentiate between the task description and date and time information.

Some of the methods in Duke class are changed to private because they modify the list which is a private attribute. This improves the encapsulation.

The code is more abstract as the main function contains only the commands. This is done by creating two new functions processCommand() and addTask().