Add save and clearsave commands
Added the save command which saves the data into a separate text file.
Previously this was only done when the "bye" command was used to end
the program.
Added the clearsave command, which clears the saved data.
Renamed getDataFromFile() in List class to addEntryFromFile() as this
more accurately reflects the functionality
Removed and replaced few magic ints related to loading saved data.
Removed dukeData.txt from being tracked by git.