Fix file saving for JAR files
Fixed the bug where the program would not save the data if run from a
JAR file in a folder with no "Data" directory. Now the "Data" directory
will be created if it does not already exist.
Added a check if there is existing data when the program starts. If no
data exists, no message will be printed. Previously, as the program
tried to load data anyway, it would not be able to find a save file and
would print an error message.