Storing the tasks in the text file
The task files are now stored in the text file so that they can still be accessed in future program runs.

There is a new function called "clear" which removes all the tasks from array and in the file. This makes it more efficient than removing every task using "delete" command.

Task has become an abstract class because it is easier to implement the new function called "toFileFormat".