Refactored the code into many classes
New packages have been created to split class based on their purpose.

'undo' command is introduced in this commit to uncheck the task. For instance, if the user has marked the wrong task as "done", this command can be used to revert it.

Newer exceptions have been created to print different message for different errors easier.